+++ src/main/Utilities.cpp
@@ -11,7 +11,7 @@
 *
 ***********************/
 
-#include "Utilities.h"
+#include "include/Utilities.h"
 
 template <typename Clientele, typename Inventory>
 int Utilities::makeTransaction(const Clientele& clientele, const Inventory& inventory, int custID) {
